feat: add 002 bogo order of bar items
